[mock] organize folder term response
