+++ src/main/java/com/example/capstone_project/controller/TermController.java
@@ -4,9 +4,9 @@ import com.example.capstone_project.controller.body.term.create.CreateTermBody;
 import com.example.capstone_project.controller.body.term.delete.DeleteTermBody;
 import com.example.capstone_project.controller.body.term.update.UpdateTermBody;
 import com.example.capstone_project.controller.responses.term.get_plans.PlanStatusResponse;
-import com.example.capstone_project.controller.responses.term.get_plans.TermDetailResponse;
+import com.example.capstone_project.controller.responses.term.get_termDetail.TermDetailResponse;
 import com.example.capstone_project.controller.responses.term.get_plans.TermPlanDetailResponse;
-import com.example.capstone_project.controller.responses.term.get_plans.TermStatusResponse;
+import com.example.capstone_project.controller.responses.term.get_termDetail.TermStatusResponse;
 import com.example.capstone_project.entity.TermDuration;
 import com.example.capstone_project.service.term.TermService;
 import com.example.capstone_project.utils.enums.TermCode;
