+++ src/main/java/com/example/capstone_project/controller/responses/term/get_termDetail/TermDetailResponse.java
@@ -1,4 +1,4 @@
-package com.example.capstone_project.controller.responses.term.get_plans;
+package com.example.capstone_project.controller.responses.term.get_termDetail;
 
 import com.example.capstone_project.entity.TermDuration;
 import jakarta.persistence.*;
