+++ src/main/java/com/example/capstone_project/service/term/TermService.java
@@ -1,10 +1,5 @@
 package com.example.capstone_project.service.term;
 
-import com.example.capstone_project.controller.responses.term.get_plans.TermDetailResponse;
-import com.example.capstone_project.controller.responses.term.get_plans.TermPlanDetailResponse;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
 public interface TermService {
