+++ src/main/java/com/example/capstone_project/utils/mapper/term/update/UpdateTermBodyToTermDetailResponseMapper.java
@@ -1,11 +1,7 @@
 package com.example.capstone_project.utils.mapper.term.update;
 
-import com.example.capstone_project.controller.body.term.delete.DeleteTermBody;
 import com.example.capstone_project.controller.body.term.update.UpdateTermBody;
-import com.example.capstone_project.controller.responses.term.get_plans.TermDetailResponse;
-import com.example.capstone_project.controller.responses.user.create.UserDetailResponse;
-import com.example.capstone_project.entity.Term;
-import com.example.capstone_project.entity.User;
+import com.example.capstone_project.controller.responses.term.get_termDetail.TermDetailResponse;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
